Refactor shared middleware code out into a module
